Move 'post methods from RLComponenetBundle to EnvSampler
